Refactor context block execution
