+++ busted/core.lua
@@ -234,81 +234,6 @@ return function()
     busted.export(alias, publisher)
   end
 
-  function busted.reject(descriptor, element)
-    local env = getfenv(element.run)
-    if env[descriptor] then
-      element.env[descriptor] = function(...)
-        error("'" .. descriptor .. "' not supported inside current context block", 2)
-      end
-    end
-  end
-
-  function busted.rejectAll(element)
-    busted.reject('randomize', element)
-    for descriptor, _ in pairs(busted.executors) do
-      busted.reject(descriptor, element)
-    end
-  end
-
-  local function exec(descriptor, element)
-    if not element.env then element.env = {} end
-
-    busted.rejectAll(element)
-
-    local parent = busted.context.parent(element)
-    setmetatable(element.env, {
-      __newindex = function(self, key, value)
-        if not parent.env then parent.env = {} end
-        parent.env[key] = value
-      end
-    })
-
-    local ret = { busted.safe(descriptor, element.run, element) }
-    return unpack(ret)
-  end
-
-  function busted.execAll(descriptor, current, propagate, err)
-    local parent = busted.context.parent(current)
-
-    if propagate and parent then
-      local success, ancestor = busted.execAll(descriptor, parent, propagate)
-      if not success then
-        return success, ancestor
-      end
-    end
-
-    local list = current[descriptor] or {}
-
-    local success = true
-    for _, v in pairs(list) do
-      if not exec(descriptor, v):success() then
-        if err then err(descriptor) end
-        success = nil
-      end
-    end
-    return success, current
-  end
-
-  function busted.dexecAll(descriptor, current, propagate, err)
-    local parent = busted.context.parent(current)
-    local list = current[descriptor] or {}
-
-    local success = true
-    for _, v in pairs(list) do
-      if not exec(descriptor, v):success() then
-        if err then err(descriptor) end
-        success = nil
-      end
-    end
-
-    if propagate and parent then
-      if not busted.dexecAll(descriptor, parent, propagate) then
-        success = nil
-      end
-    end
-    return success
-  end
-
   function busted.execute(current)
     if not current then current = busted.context.get() end
     for _, v in pairs(busted.context.children(current)) do
